+++ module/item/ritual-sheet.js
@@ -1,4 +1,6 @@
-export default class UARitualSheet extends ItemSheet
+import UABaseItemSheet from "./base-item-sheet.js";
+
+export default class UARitualSheet extends UABaseItemSheet
 {
     static get defaultOptions() {
         return mergeObject(super.defaultOptions, {
@@ -8,21 +10,8 @@ export default class UARitualSheet extends ItemSheet
                 "ritual"
             ],
             height: 600,
-            tabs: [{
-                navSelector: ".tab-buttons",
-                contentSelector: ".tab-panels",
-                initial: "main"
-            }],
             width: 650
         });
-    }
-
-    get template() {
-        if (game.user.isGM || !this.item.limited) {
-            return "systems/unknownarmies/template/item/ritual-sheet.hbs";
-        } else {
-            return "systems/unknownarmies/template/item/limited-ritual-sheet.hbs";
-        }
     }
 
     async getData (options) {
@@ -33,36 +22,6 @@ export default class UARitualSheet extends ItemSheet
         data.enrichedEffect = await TextEditor.enrichHTML(this.object.system.effect, {
             async: true
         });
-        data.enrichedPublicNotes = await TextEditor.enrichHTML(this.object.system.notes.public, {
-            async: true
-        });
-        data.enrichedPrivateNotes = await TextEditor.enrichHTML(this.object.system.notes.private, {
-            async: true
-        });
         return data;
-    }
-
-    activateListeners (html) {
-        super.activateListeners(html);
-        html.find("input").on("keydown", this._onInputKeydown.bind(this));
-        let showImage = html.find("[data-action='show-image']");
-        showImage.on("click", this._onShowImage.bind(this));
-        showImage.prop("disabled", false);
-        html.find(".editor-content--extra-small").parent().addClass("editor--extra-small");
-    }
-
-    _onInputKeydown (event) {
-        if (event.keyCode === 13) {
-            event.preventDefault();
-            super.submit();
-            $(event.currentTarget)[0].blur();
-        }
-    }
-
-    _onShowImage (event) {
-        event.preventDefault();
-        new ImagePopout(this.item.img, {
-            title: this.item.name
-        }).render(true);
     }
 }